requirements.txt
requests
pyyaml
plexapi
sonarr-py
lxml
python-dotenv